fix: create separate component for loading info banner
